Move data structures to different package
